+++ packages/eslint-plugin-isaacscript/tests/rules/enum-member-number-separation.test.ts
@@ -12,6 +12,7 @@ enum Foo {
       `,
     },
   ],
+
   invalid: [
     {
       code: `