+++ packages/eslint-plugin-isaacscript/tests/rules/eqeqeq-fix.test.ts
@@ -16,6 +16,7 @@ if (foo !== bar) {
       `,
     },
   ],
+
   invalid: [
     {
       code: `