+++ packages/eslint-plugin-isaacscript/tests/rules/jsdoc-code-block-language.test.ts
@@ -37,6 +37,7 @@ ruleTester.run("jsdoc-code-block-language", jsdocCodeBlockLanguage, {
       `,
     },
   ],
+
   invalid: [
     {
       code: `