+++ packages/eslint-plugin-isaacscript/tests/rules/no-empty-jsdoc.test.ts
@@ -16,6 +16,7 @@ ruleTester.run("no-empty-jsdoc", noEmptyJSDoc, {
       `,
     },
   ],
+
   invalid: [
     {
       code: `