+++ packages/eslint-plugin-isaacscript/tests/rules/no-implicit-map-set-loops.test.ts
@@ -16,6 +16,7 @@ for (const value of mySet.values()) {}
       `,
     },
   ],
+
   invalid: [
     {
       code: `