+++ packages/eslint-plugin-isaacscript/tests/rules/no-let-any.test.ts
@@ -20,6 +20,7 @@ let [, secondElement] = myArray;
       `,
     },
   ],
+
   invalid: [
     {
       code: `