+++ packages/eslint-plugin-isaacscript/tests/rules/no-object-any.test.ts
@@ -25,6 +25,7 @@ const mySet: Set<string> = new Set();
       `,
     },
   ],
+
   invalid: [
     {
       code: `