+++ packages/eslint-plugin-isaacscript/tests/rules/no-template-curly-in-string-fix.test.ts
@@ -9,6 +9,7 @@ const fooString = \`foo: \${foo}\`;
       `,
     },
   ],
+
   invalid: [
     {
       code: `