+++ packages/eslint-plugin-isaacscript/tests/rules/no-useless-return-no-fix.test.ts
@@ -11,6 +11,7 @@ function foo() {
       `,
     },
   ],
+
   invalid: [
     {
       code: `