+++ packages/eslint-plugin-isaacscript/tests/rules/no-void-return-type.test.ts
@@ -9,6 +9,7 @@ function foo() {}
       `,
     },
   ],
+
   invalid: [
     {
       code: `